Let e2e barely execute by excluding the commons-net jar package.
