+++ seatunnel-connectors-v2/connector-file/connector-file-base/pom.xml
@@ -30,7 +30,6 @@
     <name>SeaTunnel : Connectors V2 : File : Base</name>
 
     <properties>
-        <commons-net.version>3.6</commons-net.version>
         <orc.version>1.5.6</orc.version>
         <commons.collecton4.version>4.4</commons.collecton4.version>
         <commons.lang3.version>3.4</commons.lang3.version>
@@ -81,12 +80,6 @@
             <groupId>org.apache.parquet</groupId>
             <artifactId>parquet-avro</artifactId>
             <version>${parquet-avro.version}</version>
-        </dependency>
-
-        <dependency>
-            <groupId>commons-net</groupId>
-            <artifactId>commons-net</artifactId>
-            <version>${commons-net.version}</version>
         </dependency>
 
         <dependency>
