+++ seatunnel-e2e/seatunnel-connector-v2-e2e/connector-file-ftp-e2e/pom.xml
@@ -22,7 +22,7 @@
         <version>${revision}</version>
     </parent>
 
-    <artifactId>connector-file-local-e2e</artifactId>
+    <artifactId>connector-file-ftp-e2e</artifactId>
     <name>SeaTunnel : E2E : Connector V2 : File Ftp</name>
 
     <dependencies>
