+++ seatunnel-e2e/seatunnel-connector-v2-e2e/connector-file-ftp-e2e/src/test/resources/excel/fake_source_to_ftp_excel.conf
@@ -30,7 +30,7 @@ env {
 
 source {
   FakeSource {
-    result_table_name = "fake"
+    result_table_name = "ftp"
     schema = {
       fields {
         c_map = "map<string, string>"
@@ -70,12 +70,17 @@ source {
 
 sink {
   FtpFile {
-    path="/tmp/seatunnel/excel"
-    partition_dir_expression="${k0}=${v0}"
-    is_partition_field_write_in_file=true
-    file_name_expression="${transactionId}_${now}"
-    file_format_type="excel"
-    filename_time_format="yyyy.MM.dd"
-    is_enable_transaction=true
+    host = "ftp"
+    port = 21
+    user = seatunnel
+    password = pass
+    path = "/"
+    source_table_name = "ftp"
+    partition_dir_expression = "${k0}=${v0}"
+    is_partition_field_write_in_file = true
+    file_name_expression = "${transactionId}_${now}"
+    file_format_type = "excel"
+    filename_time_format = "yyyy.MM.dd"
+    is_enable_transaction = true
   }
 }
