+++ seatunnel-e2e/seatunnel-connector-v2-e2e/connector-file-ftp-e2e/src/test/resources/excel/ftp_excel_projection_to_assert.conf
@@ -20,14 +20,33 @@
 
 source {
   FtpFile {
-    path="/seatunnel/read/excel"
-    result_table_name = "fake"
+    host = "ftp"
+    port = 21
+    user = seatunnel
+    password = pass
+    path = "/seatunnel/read/excel"
+    result_table_name = "ftp"
     file_format_type = excel
     delimiter = ;
     read_columns = [c_string, c_boolean]
     skip_header_row_number = 1
     schema = {
       fields {
+        c_map = "map<string, string>"
+        c_array = "array<int>"
+        c_string = string
+        c_boolean = boolean
+        c_tinyint = tinyint
+        c_smallint = smallint
+        c_int = int
+        c_bigint = bigint
+        c_float = float
+        c_double = double
+        c_bytes = bytes
+        c_date = date
+        c_decimal = "decimal(38, 18)"
+        c_timestamp = timestamp
+        c_row = {
           c_map = "map<string, string>"
           c_array = "array<int>"
           c_string = string
@@ -42,22 +61,7 @@ source {
           c_date = date
           c_decimal = "decimal(38, 18)"
           c_timestamp = timestamp
-          c_row = {
-            c_map = "map<string, string>"
-            c_array = "array<int>"
-            c_string = string
-            c_boolean = boolean
-            c_tinyint = tinyint
-            c_smallint = smallint
-            c_int = int
-            c_bigint = bigint
-            c_float = float
-            c_double = double
-            c_bytes = bytes
-            c_date = date
-            c_decimal = "decimal(38, 18)"
-            c_timestamp = timestamp
-          }
+        }
       }
     }
   }
@@ -66,6 +70,7 @@ source {
 
 sink {
   Assert {
+    source_table_name = "ftp"
     rules {
       row_rules = [
         {
