+++ seatunnel-e2e/seatunnel-connector-v2-e2e/connector-file-ftp-e2e/src/test/resources/json/fake_to_ftp_file_json.conf
@@ -27,6 +27,7 @@ env {
 
 source {
   FtpFile {
+    result_table_name = "ftp"
     schema = {
       fields {
         c_map = "map<string, string>"
@@ -61,13 +62,17 @@ source {
         }
       }
     }
-    result_table_name = "fake"
   }
 }
 
 sink {
-  LocalFile {
-    path = "/tmp/seatunnel/json"
+  FtpFile {
+    host = "ftp"
+    port = 21
+    user = seatunnel
+    password = pass
+    path = "/seatunnel/json"
+    source_table_name = "ftp"
     row_delimiter = "\n"
     partition_dir_expression = "${k0}=${v0}"
     is_partition_field_write_in_file = true
