+++ seatunnel-e2e/seatunnel-connector-v2-e2e/connector-file-ftp-e2e/src/test/resources/text/ftp_file_text_projection_to_assert.conf
@@ -27,9 +27,14 @@ env {
 
 source {
   FtpFile {
+    host = "ftp"
+    port = 21
+    user = seatunnel
+    password = pass
     path = "/seatunnel/read/text"
     file_format_type = "text"
     read_columns = [c_string, c_boolean, c_double]
+    result_table_name = "ftp"
     schema = {
       fields {
         c_map = "map<string, string>"
@@ -64,12 +69,12 @@ source {
         }
       }
     }
-    result_table_name = "fake"
   }
 }
 
 sink {
   Assert {
+    source_table_name = "ftp"
     rules {
       row_rules = [
         {
