+++ seatunnel-e2e/seatunnel-connector-v2-e2e/connector-file-ftp-e2e/src/test/resources/text/ftp_file_text_to_assert.conf
@@ -27,8 +27,13 @@ env {
 
 source {
   FtpFile {
+    host = "ftp"
+    port = 21
+    user = seatunnel
+    password = pass
     path = "/seatunnel/read/text"
     file_format_type = "text"
+    result_table_name = "ftp"
     schema = {
       fields {
         c_map = "map<string, string>"
@@ -63,12 +68,12 @@ source {
         }
       }
     }
-    result_table_name = "fake"
   }
 }
 
 sink {
   Assert {
+    source_table_name = "ftp"
     rules {
       row_rules = [
         {
